On the 64-bit SGI Origin2000, you can't cast from a pointer to an integer ("conversion is invalid"): 	int i=(int)ptr; This means MPI_Aint should probably be a "long" (a 64-bit int) rather than just an "int" (32-bit int), but doing so would break compatability between f90 and C, so we'll leave AMPI_Aint as "int" and add the horriffic triple cast: 	int i=(int)(long)(char *)ptr;

+++ src/libs/ck-libs/ampi/ampi.C
@@ -2212,7 +2212,7 @@ int MPI_Type_ub(MPI_Datatype dtype, MPI_Aint* displacement){
 CDECL
 int MPI_Address(void* location, MPI_Aint *address){
   AMPIAPI("MPI_Address");
-  *address = (MPI_Aint)((char *)location);
+  *address = (MPI_Aint)(unsigned long)(char *)location;
   return 0;
 }
